#6 when the Docker container fails to run, the Spring container is killed
Also fixed an error in the Docker process runner. The error was that the Docker command has two port references, mapping the local port to the VM port. The variable part was put on the VM port which is (currently) not settable from the outside. The local port should be. The two values have been switched, so now it is possible to run two Postgres containers in parallel.

Signed-off-by: robert-bor <bor.robert@gmail.com>